fix: typo and rename JiraComponent to IssueTrackerComponent
