+++ frontend/src/components/StoryPointsComponent.vue
